Reject nonregular released artifacts before hashing and share digest helper
